feat: development plugin for support linux-amd64 architecture

+++ lib/utils.bash
@@ -2,7 +2,6 @@
 
 set -euo pipefail
 
-# TODO: Ensure this is the correct GitHub homepage where releases can be downloaded for argocd-autopilot.
 GH_REPO="https://github.com/argoproj-labs/argocd-autopilot"
 TOOL_NAME="argocd-autopilot"
 TOOL_TEST="argocd-autopilot version"
@@ -31,8 +30,6 @@ list_github_tags() {
 }
 
 list_all_versions() {
-	# TODO: Adapt this. By default we simply list the tag names from GitHub releases.
-	# Change this function if argocd-autopilot has other means of determining installable versions.
 	list_github_tags
 }
 
@@ -41,11 +38,15 @@ download_release() {
 	version="$1"
 	filename="$2"
 
-	# TODO: Adapt the release URL convention for argocd-autopilot
-	url="$GH_REPO/archive/v${version}.tar.gz"
+	url="$GH_REPO/releases/download/v${version}/argocd-autopilot-linux-amd64.tar.gz"
 
-	echo "* Downloading $TOOL_NAME release $version..."
-	curl "${curl_opts[@]}" -o "$filename" -C - "$url" || fail "Could not download $url"
+	if [ ! -f "$filename" ]; then
+		echo "* Downloading $TOOL_NAME release $version..."
+		curl "${curl_opts[@]}" -o "$filename" -C - "$url" || fail "Could not download $url"
+	else
+		echo "$TOOL_NAME $version already exists. No download needed."
+	fi
+
 }
 
 install_version() {
@@ -61,7 +62,6 @@ install_version() {
 		mkdir -p "$install_path"
 		cp -r "$ASDF_DOWNLOAD_PATH"/* "$install_path"
 
-		# TODO: Assert argocd-autopilot executable exists.
 		local tool_cmd
 		tool_cmd="$(echo "$TOOL_TEST" | cut -d' ' -f1)"
 		test -x "$install_path/$tool_cmd" || fail "Expected $install_path/$tool_cmd to be executable."
